+++ docs/user/changes.rst
@@ -9,6 +9,8 @@ Release history
 - feat: new method :meth:`api.converged_recordings.ConvergedRecordingsApi.restore_recordings_from_recycle_bin <wxc_sdk.converged_recordings.ConvergedRecordingsApi.restore_recordings_from_recycle_bin>`
 - feat: new method :meth:`api.converged_recordings.ConvergedRecordingsApi.purge_recordings_from_recycle_bin <wxc_sdk.converged_recordings.ConvergedRecordingsApi.purge_recordings_from_recycle_bin>`
 - fix: :meth:`api.devices.create_by_mac_address <wxc_sdk.devices.DevicesApi.create_by_mac_address>` returns None if device creation leads to empty response. Apparently this is the case for deskphone devices
+- feat: new guest calling settings API: :attr:`api.telephony.guest_calling <wxc_sdk.telephony.TelephonyApi.guest_calling>`
+- feat: new endpoint to get call token for click-to-call :meth:`api.telephony.create_a_call_token <wxc_sdk.telephony.TelephonyApi.create_a_call_token>`
 
 1.23.0
 ======